added assets and moved mockcats.js, removed navlink from header, imported correct array to app.js
