+++ src/assets/MockCats.js
