+++ src/pages/Home.js
@@ -1,4 +1,5 @@
 import React from 'react'
+import mockCats from "./assets/MockCats"
 
 const Home = () => {
     return (
